bring the env back to prefix metrics
this is because we're going to use our production librato account
so it's important that each telex instance instruments on a diff
namespace.

using console_banner here because rack_env is production for both
prod and staging. Ideally I think Pliny should just call this
"deployment" instead, will start a thread on this.