Re-raise HTTPStatusErrors so Pliny middleware can handle it
Pliny::Middleware::RescueErrors properly cleans env["sinatra.error"] so
the Rollbar middleware won't report on 404 errors.
We have to set the status in Base, because
Pliny::Extensions::Instruments reports it as 500 once it leaves the
"App" and into middleware (－‸ლ)

Updating to the latest Pliny (a larger task) may solve some of these
problems, the instrumentation is completely rewritten.